+++ modules/gui/widgets/realslider.h
@@ -25,7 +25,7 @@ typedef std::shared_ptr<OtRealSliderClass> OtRealSlider;
 class OtRealSliderClass : public OtWidgetClass {
 public:
 	// initialize slider
-	void init(const std::string& label, double min, double max, double value, OtObject callback);
+	void init(const std::string& label, float min, float max, float value, OtObject callback);
 
 	// render content
 	void render();
