+++ modules/gui/widgets/screen.h
@@ -31,10 +31,10 @@ public:
 	size_t getMenubarHeight();
 
 	// mouse and keyboard events
-	void onMouseButton(int button, int action, int mods, double xpos, double ypos);
-	void onMouseMove(double xpos, double ypos);
-	void onMouseDrag(int button, int mods, double xpos, double ypos);
-	void onScrollWheel(double dx, double dy);
+	void onMouseButton(int button, int action, int mods, float xpos, float ypos);
+	void onMouseMove(float xpos, float ypos);
+	void onMouseDrag(int button, int mods, float xpos, float ypos);
+	void onScrollWheel(float dx, float dy);
 	void onKey(int key, int mods);
 	void onChar(unsigned int codepoint);
